Safari-609.1.11.1/LayoutTests/editing/selection/contains-node-crash-expected.txt

ALERT: PASS: containsNode returned false
Test calling containsNode on a selection of a closed document doesn't cause a crash.
